requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
openai
tiktoken
